+++ packages/dnb-design-system-portal/src/pages/uilib/components/date-picker/Examples.js
@@ -5,22 +5,16 @@
 
 import React, { PureComponent, Fragment } from 'react'
 import ComponentBox from '../../../../shared/tags/ComponentBox'
-import DatePickerRange from 'dnb-ui-lib/src/components/date-picker/DatePickerRange'
+import styled from '@emotion/styled'
 
 class Example extends PureComponent {
   render() {
     const IS_TEST = typeof window !== 'undefined' && window.IS_TEST
+    if (IS_TEST) {
+      return <ScreenshotTests />
+    }
     return (
       <Fragment>
-        <ComponentBox data-dnb-test={IS_TEST && 'date-picker-input'}>
-          {/* @jsx */ `
-<DatePicker
-  label="Range DatePicker:"
-  range={true}
-  show_input={true}
-/>
-          `}
-        </ComponentBox>
         <ComponentBox>
           {/* @jsx */ `
 <DatePicker
@@ -44,7 +38,7 @@ class Example extends PureComponent {
         <ComponentBox>
           {/* @jsx */ `
 <DatePicker
-  label="Defualt DatePicker with Input:"
+  label="Default DatePicker with Input:"
   date={new Date()}
   show_input={true}
   show_cancel_button={true}
@@ -53,21 +47,6 @@ class Example extends PureComponent {
   }}
   on_cancel={({ date }) => {
     console.log('on_cancel', date)
-  }}
-/>
-          `}
-        </ComponentBox>
-        <ComponentBox>
-          {/* @jsx */ `
-<DatePicker
-  label="Defualt DatePicker:"
-  date="2019-05-05"
-  return_format="DD-MM-YYYY"
-  on_change={({ date }) => {
-    console.log('on_change', date)
-  }}
-  on_show={({ date }) => {
-    console.log('on_show', date)
   }}
 />
           `}
@@ -98,24 +77,109 @@ class Example extends PureComponent {
 />
           `}
         </ComponentBox>
+        <ComponentBox>
+          {/* @jsx */ `
+<DatePicker
+  label="DatePicker with error message:"
+  date={new Date()}
+  show_input={true}
+  show_submit_button={true}
+  status="Please select a valid date"
+/>
+          `}
+        </ComponentBox>
+        <ComponentBox>
+          {/* @jsx */ `
+<DatePicker
+  label="DatePicker with error status:"
+  date={new Date()}
+  hide_navigation={true}
+  status="error"
+/>
+          `}
+        </ComponentBox>
+        <ComponentBox>
+          {/* @jsx */ `
+<DatePicker
+  label="Disabled with info message:"
+  date={new Date()}
+  show_input={true}
+  status="Please select a valid date"
+  status_state="info"
+  disabled
+/>
+          `}
+        </ComponentBox>
 
-        <div className="example-box">
-          <div data-dnb-test="date-picker-calendar">
-            <DatePickerRange
-              range={true}
-              startDate={new Date('2019-05-05')}
-              endDate={new Date('2019-06-05')}
-              firstDayOfWeek={'monday'}
-            />
-          </div>
-          <p className="example-caption">
-            Example styling of range calendar (also used for screenshot
-            tests)
-          </p>
-        </div>
+        <ScreenshotTests />
       </Fragment>
     )
   }
 }
 
 export default Example
+
+const ScreenshotTests = () => {
+  return (
+    <Wrapper>
+      <ComponentBox data-dnb-test="date-picker-trigger-default">
+        {/* @jsx */ `
+<DatePicker
+  label="Default DatePicker:"
+  date="2019-05-05"
+  return_format="DD-MM-YYYY"
+  on_change={({ date }) => {
+    console.log('on_change', date)
+  }}
+  on_show={({ date }) => {
+    console.log('on_show', date)
+  }}
+/>
+        `}
+      </ComponentBox>
+      <ComponentBox data-dnb-test="date-picker-input">
+        {/* @jsx */ `
+<DatePicker
+  label="Range DatePicker:"
+  range={true}
+  show_input={true}
+/>
+        `}
+      </ComponentBox>
+      <ComponentBox data-dnb-test="date-picker-trigger-error">
+        {/* @jsx */ `
+<DatePicker
+  label="DatePicker with error status:"
+  date={new Date()}
+  hide_navigation={true}
+  status="Please select a valid date"
+/>
+    `}
+      </ComponentBox>
+      <ComponentBox
+        data-dnb-test="date-picker-calendar"
+        caption="Opened DatePicker (also used for screenshot tests)"
+      >
+        {/* @jsx */ `
+<DatePicker
+  opened="true"
+  range="true"
+  start_date="2019-05-05"
+  end_date="2019-06-05"
+/>
+    `}
+      </ComponentBox>
+    </Wrapper>
+  )
+}
+
+const Wrapper = styled.div`
+  [data-dnb-test='date-picker-calendar'] .dnb-date-picker__container {
+    display: block;
+    position: relative;
+    top: 0;
+  }
+  [data-dnb-test='date-picker-calendar'] .dnb-date-picker {
+    margin-left: 1rem;
+  }
+`