+++ packages/dnb-design-system-portal/src/pages/uilib/usage/customisation/styling/legacy-styling.js
@@ -71,7 +71,7 @@ const LegacyCodeStyling = () => (
     <SpacingExample className="dnb-core-style dnb-spacing dnb-dev-grid">
       <Code>{'<div class="dnb-core-style dnb-spacing">'}</Code>
       <H2>
-        I'm a <Code>{`<h2>`}</Code> in an Article, and have the defualt
+        I'm a <Code>{`<h2>`}</Code> in an Article, and have the default
         spacing!
       </H2>
       <P>