+++ packages/dnb-ui-lib/src/components/date-picker/__tests__/DatePicker.screenshot.test.js
@@ -26,4 +26,18 @@ describe('DatePicker screenshot', () => {
     })
     expect(screenshot).toMatchImageSnapshot()
   })
+  it('have to match the date-picker trigger button', async () => {
+    const screenshot = await testPageScreenshot({
+      selector:
+        '[data-dnb-test="date-picker-trigger-default"] .dnb-date-picker'
+    })
+    expect(screenshot).toMatchImageSnapshot()
+  })
+  it('have to match the date-picker trigger button in error state', async () => {
+    const screenshot = await testPageScreenshot({
+      selector:
+        '[data-dnb-test="date-picker-trigger-error"] .dnb-date-picker'
+    })
+    expect(screenshot).toMatchImageSnapshot()
+  })
 })