+++ CHANGELOG.md
@@ -1,10 +1,14 @@
-## May, 27. 2019
+## June, 5. 2019
+
+- New corner radius (border-radius) on `Dropdown` and `DatePicker`: 4px (0.25rem)
+
+## June, 3. 2019
 
 - New component in the `dnb-ui-lib`: [Textarea](/uilib/components/textarea). Got also a wider corner radius by `16px (1rem)`.
 
 ## May, 27. 2019
 
-- New component in the `dnb-ui-lib`: [Checkbox](/uilib/components/checkbox) and [Radio](/uilib/components/radio) button
+- New components in the `dnb-ui-lib`: [Checkbox](/uilib/components/checkbox) and [Radio](/uilib/components/radio) button
 
 ## May, 19. 2019